chore(devops): Update how application handle build stages
