--- tools/webpack.server.js
@@ -1,49 +1,0 @@
-const webpack = require('webpack')
-const fs = require('fs')
-const path = require('path')
-
-function getExternals () {
-  const nodeModules = fs.readdirSync(path.join(process.cwd(), 'node_modules'))
-  return nodeModules.reduce(function (ext, mod) {
-    ext[mod] = 'commonjs ' + mod
-    return ext
-  }, {})
-}
-module.exports = {
-  target: 'node',
-  devtool: 'nosources-source-map',
-  entry: './server',
-  output: {
-    path: path.join(__dirname, './../build'),
-    filename: 'server.js'
-  },
-  externals: getExternals(),
-  node: {
-    __filename: true,
-    __dirname: true
-  },
-  resolve: {
-    alias: {
-      './rangy-core': 'rangy/lib/rangy-core'
-    }
-  },
-  module: {
-    loaders: [
-      {
-        test: /\.js$/,
-        loader: 'babel-loader',
-        exclude: /(node_modules|wysihtml-toolbar.min)/
-      }
-    ]
-  },
-  plugins: [
-    new webpack.BannerPlugin(
-      {
-        banner: 'require("source-map-support").install();',
-        raw: true,
-        entryOnly: false
-      }
-    ),
-    new webpack.IgnorePlugin(/\.(css|less|scss|svg|png|jpe?g|png)$/)
-  ]
-}
